Updated git with Database_Store.php Began creating intermediate datalayer between frontend and backend 3/16/22 Updated test file to reflect the new Line Item factory 3/17/22 Updated test file and Line Item factory with initial validation logic 3/17/22 Tweaked the testing file formatting and minor changes to the  Line Item factory 3/17/22 After a few days of fighting with Apache & MYSQL had to rebuild Database_Store 3/24/22 Further adventures in SQL errors! 3/28/22 added ADT for strings and numbers added further data types cleaning up after the new types were introduced removing composer vendor folder from git updated Tax and Taxes classes More work on the Tax and Taxes classes. Ready to test and work the kinks out.

+++ src/Currency_Field.php
@@ -8,12 +8,12 @@ class Currency_Field{
     public static array $error = [];
 
 
-    public function __construct(float $number){
+    public function __construct(int $number){
         self::get_errors($number);
         $this->item_currency = $number;
     }
 
-        public function currency(): float{
+        public function currency(): int{
         return $this->item_currency;
     }
